refactor: extract UI from app.py into src/ui/ package (Phase 16.12)
Decompose 1,861-line app.py into focused UI modules:
- ui/styles.py: CSS, page config, logo encoding
- ui/lock.py: device lock UI (read-only mode, banners, takeover)
- ui/chat.py: chat interface and streaming
- ui/panels.py: 14 display components (transparency, patterns, etc.)
- ui/network.py: 5 trusted network components (setup, handoff, etc.)
- ui/sidebar.py: sidebar rendering and session management
- ui/__init__.py: public API

app.py is now a 233-line thin orchestrator.
All 796 tests pass with no changes to test files.
